+++ fe2o3-amqp/src/transaction/controller.rs
@@ -64,9 +64,9 @@ where
 impl Controller {
     /// Creates a new builder for controller
     pub fn builder(
-    ) -> link::builder::Builder<role::Sender, Coordinator, WithoutName, WithSource, WithoutTarget>
+    ) -> link::builder::Builder<role::SenderMarker, Coordinator, WithoutName, WithSource, WithoutTarget>
     {
-        link::builder::Builder::<role::Sender, Coordinator, WithoutName, WithSource, WithoutTarget>::new()
+        link::builder::Builder::<role::SenderMarker, Coordinator, WithoutName, WithSource, WithoutTarget>::new()
     }
 
     /// Close the control link with error